requirements.txt
openai
panel
watchfiles
llama-index-agent-openai
llama-index-llms-openai